fast_vdf: Optimize FastPowFormNucomp
Do not compute form squaring on the last iteration when its result is
not used.

Do not reduce form after every composition or squaring, but only when
the result's 'a' coefficient exceeds the discriminant in size. This is
efficient because nucomp produces forms that are 'nearly reduced', so
their coefficients' sizes are usually below discriminant size, and we
can do reductions much less often.